pyproject.toml
[project]
name = "teams-productivity-bot"
description = "AI-Powered Microsoft Teams Productivity Bot with advanced features"
requires-python = ">=3.11"
version = "2.0.0"

dependencies = [
  "flask>=2.3.0",
  "botbuilder-core>=4.15.0",
  "botbuilder-schema>=4.15.0",
  "requests>=2.31.0",
  "aiohttp>=3.8.0",
  "python-dotenv>=1.0.0",
  "gunicorn>=21.0.0",
  "qrcode[pil]>=7.4.2",
  "pillow>=10.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0.0",
  "pytest-asyncio>=0.21.0",
  "black>=23.0.0",
  "flake8>=6.0.0"
]

[tool.uv]
